수정된 도메인에 따른 controller 변경
